+++ src/PPN-microbench/microbench.cpp
@@ -1,6 +1,6 @@
 #include <PPN-microbench/microbench.hpp>
 
-Microbench::Microbench(std::string name, int nbIterations) {
+Microbench::Microbench(std::string name, uint64_t nbIterations) {
     this->name = name;
     this->nbIterations = nbIterations;
 }
@@ -9,4 +9,4 @@ Microbench::~Microbench() {}
 
 std::string Microbench::getName() { return this->name; }
 
-int Microbench::getNbIterations() { return this->nbIterations; }
+uint64_t Microbench::getNbIterations() { return this->nbIterations; }
